feat: add Discord API models for webhooks, forum posts, and related entities. Also, beginning of the bulletin board system

+++ starbound-protocol/src/main/java/irden/space/proxy/protocol/codec/variant/IntVariantValue.java
@@ -1,4 +1,4 @@
 package irden.space.proxy.protocol.codec.variant;
 
-public record IntVariantValue(int value) implements VariantValue {
+public record IntVariantValue(long value) implements VariantValue {
 }
